Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/backend/requirements.txt
Flask
markupsafe==2.0.1
requests
pylint==2.4.0
gunicorn==20.0.4
pymongo==3.11.0
Flask-PyMongo==2.3.0
prometheus-client==0.12.0
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation
opentelemetry-sdk
prometheus_flask_exporter
Flask-Opentracing
